Pull request update for #11 (#12)
* fix: docker container not listening on all interfaces

The environment variable in the Dockerfile needed to set SERVER__HOST to 0.0.0.0 to listen on all interfaces.

PORT was being set, but it wasn't making its way into the server, causing it to listen on only the loopback address by default.

This was preventing access from the docker host even if the port was exposed.

* fix: update Docker configs to use consistent SERVER__ env vars

---------

Co-authored-by: Dannel Albert <cuebix@gmail.com>